Refactor AST test utilities and tests
Moved AstFeature, SupportedAstFeatures, MaximalAstGenerator, and LanguageConfig to a new AstTestUtils.kt file. Extracted MaximalAstGeneratorTest to its own file. TranspilationTest.kt now only contains transpilation-related tests, improving test organization and maintainability.
